fix: wakuflag for ListenAddresses ENR
Also, reorganize code in utils package
